Break up flakeref.rs into submodules
