+++ src/flakeref.rs
@@ -7,10 +7,18 @@ use nom::{
 };
 use serde::{Deserialize, Serialize};
 
+
 use crate::{
     error::{NixUriError, NixUriResult},
     parser::{parse_owner_repo_ref, parse_url_type},
 };
+
+mod fr_type;
+pub use fr_type::FlakeRefType;
+mod fr_params;
+pub use fr_params::{FlakeRefParameters, FlakeRefParamKeys};
+mod fr_urls;
+pub use fr_urls::UrlType;
 
 /// The General Flake Ref Schema
 #[derive(Debug, Default, Clone, Serialize, Deserialize, PartialEq, Eq)]
@@ -62,181 +70,6 @@ impl Display for FlakeRef {
     }
 }
 
-#[derive(Debug, Default, Clone, Serialize, Deserialize, PartialEq, Eq)]
-#[cfg_attr(test, serde(deny_unknown_fields))]
-pub struct FlakeRefParameters {
-    /// The subdirectory of the flake in which flake.nix is located. This parameter
-    /// enables having multiple flakes in a repository or tarball. The default is the
-    /// root directory of the flake.
-    dir: Option<String>,
-    /// The hash of the NAR serialisation (in SRI format) of the contents of the flake.
-    /// This is useful for flake types such as tarballs that lack a unique content
-    /// identifier such as a Git commit hash.
-    #[serde(rename = "narHash")]
-    nar_hash: Option<String>,
-    /// A Git or Mercurial commit hash.
-    rev: Option<String>,
-    ///  A Git or Mercurial branch or tag name.
-    r#ref: Option<String>,
-    branch: Option<String>,
-    submodules: Option<String>,
-    shallow: Option<String>,
-    // Only available to certain types
-    host: Option<String>,
-    // Not available to user
-    #[serde(rename = "revCount")]
-    rev_count: Option<String>,
-    // Not available to user
-    #[serde(rename = "lastModified")]
-    last_modified: Option<String>,
-    /// Arbitrary uri parameters will be allowed during initial parsing
-    /// in case they should be checked for known types run `self.check()`
-    arbitrary: Vec<(String, String)>,
-}
-
-// TODO: convert into macro!
-// or have params in a vec of tuples? with param and option<string>
-impl Display for FlakeRefParameters {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        let mut res = String::new();
-        if let Some(dir) = &self.dir {
-            res.push_str("dir=");
-            res.push_str(dir);
-        }
-        if let Some(branch) = &self.branch {
-            if !res.is_empty() {
-                res.push('?');
-            }
-            res.push_str("branch=");
-            res.push_str(branch);
-        }
-        if let Some(host) = &self.host {
-            if !res.is_empty() {
-                res.push('?');
-            }
-            res.push_str("host=");
-            res.push_str(host);
-        }
-        if let Some(r#ref) = &self.r#ref {
-            if !res.is_empty() {
-                res.push('?');
-            }
-            res.push_str("ref=");
-            res.push_str(r#ref);
-        }
-        if let Some(rev) = &self.rev {
-            if !res.is_empty() {
-                res.push('?');
-            }
-            res.push_str("rev=");
-            res.push_str(rev);
-        }
-        write!(f, "{res}")
-    }
-}
-
-impl FlakeRefParameters {
-    pub fn dir(&mut self, dir: Option<String>) -> &mut Self {
-        self.dir = dir;
-        self
-    }
-
-    pub fn nar_hash(&mut self, nar_hash: Option<String>) -> &mut Self {
-        self.nar_hash = nar_hash;
-        self
-    }
-
-    pub fn host(&mut self, host: Option<String>) -> &mut Self {
-        self.host = host;
-        self
-    }
-    pub fn rev(&mut self, rev: Option<String>) -> &mut Self {
-        self.rev = rev;
-        self
-    }
-    pub fn r#ref(&mut self, r#ref: Option<String>) -> &mut Self {
-        self.r#ref = r#ref;
-        self
-    }
-
-    pub fn set_dir(&mut self, dir: Option<String>) {
-        self.dir = dir;
-    }
-
-    pub fn set_nar_hash(&mut self, nar_hash: Option<String>) {
-        self.nar_hash = nar_hash;
-    }
-
-    pub fn set_rev(&mut self, rev: Option<String>) {
-        self.rev = rev;
-    }
-
-    pub fn set_ref(&mut self, r#ref: Option<String>) {
-        self.r#ref = r#ref;
-    }
-
-    pub fn set_host(&mut self, host: Option<String>) {
-        self.host = host;
-    }
-
-    pub fn rev_count_mut(&mut self) -> &mut Option<String> {
-        &mut self.rev_count
-    }
-
-    pub fn set_branch(&mut self, branch: Option<String>) {
-        self.branch = branch;
-    }
-
-    pub fn set_submodules(&mut self, submodules: Option<String>) {
-        self.submodules = submodules;
-    }
-
-    pub fn set_shallow(&mut self, shallow: Option<String>) {
-        self.shallow = shallow;
-    }
-    pub fn add_arbitrary(&mut self, arbitrary: (String, String)) {
-        self.arbitrary.push(arbitrary);
-    }
-    pub fn get_rev(&self) -> Option<&String> {
-        self.rev.as_ref()
-    }
-    pub fn get_ref(&self) -> Option<&String> {
-        self.r#ref.as_ref()
-    }
-}
-
-pub enum FlakeRefParamKeys {
-    Dir,
-    NarHash,
-    Host,
-    Ref,
-    Rev,
-    Branch,
-    Submodules,
-    Shallow,
-    Arbitrary(String),
-}
-
-impl std::str::FromStr for FlakeRefParamKeys {
-    type Err = NixUriError;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        use FlakeRefParamKeys::*;
-        match s {
-            "dir" | "&dir" => Ok(Dir),
-            "nar_hash" | "&nar_hash" => Ok(NarHash),
-            "host" | "&host" => Ok(Host),
-            "rev" | "&rev" => Ok(Rev),
-            "ref" | "&ref" => Ok(Ref),
-            "branch" | "&branch" => Ok(Branch),
-            "submodules" | "&submodules" => Ok(Submodules),
-            "shallow" | "&shallow" => Ok(Shallow),
-            arbitrary => Ok(Arbitrary(arbitrary.into())),
-            // unknown => Err(NixUriError::UnknownUriParameter(unknown.into())),
-        }
-    }
-}
-
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
 pub enum GitForge {
     GitHub,
@@ -256,306 +89,7 @@ impl Display for GitForge {
         )
     }
 }
-#[derive(Debug, Default, Clone, Serialize, Deserialize, PartialEq, Eq)]
-#[non_exhaustive]
-pub enum FlakeRefType {
-    // In URL form, the schema must be file+http://, file+https:// or file+file://. If the extension doesn’t correspond to a known archive format (as defined by the tarball fetcher), then the file+ prefix can be dropped.
-    File {
-        url: String,
-    },
-    /// Git repositories. The location of the repository is specified by the attribute
-    /// `url`. The `ref` arrribute defaults to resolving the `HEAD` reference.
-    /// The `rev` attribute must exist in the branch or tag specified by `ref`, defaults
-    /// to `ref`.
-    Git {
-        url: String,
-        r#type: UrlType,
-    },
 
-    GitForge {
-        platform: GitForge,
-        owner: String,
-        repo: String,
-        ref_or_rev: Option<String>,
-    },
-    Indirect {
-        id: String,
-        ref_or_rev: Option<String>,
-    },
-    // Matches `git` type, but schema is one of the following:
-    // `hg+http`, `hg+https`, `hg+ssh` or `hg+file`.
-    Mercurial {
-        url: String,
-        r#type: UrlType,
-    },
-    /// Path must be a directory in the filesystem containing a `flake.nix`.
-    /// Path must be an absolute path.
-    Path {
-        path: String,
-    },
-    Tarball {
-        url: String,
-        r#type: UrlType,
-    },
-    #[default]
-    None,
-}
-
-impl Display for FlakeRefType {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            FlakeRefType::File { url } => write!(f, "file:{url}"),
-            FlakeRefType::Git { url, r#type } => {
-                if let UrlType::None = r#type {
-                    return write!(f, "git:{url}");
-                }
-                let uri = format!("git+{}:{url}", r#type);
-                write!(f, "{uri}")
-            }
-            FlakeRefType::GitForge {
-                platform,
-                owner,
-                repo,
-                ref_or_rev,
-            } => {
-                write!(f, "{platform}:{owner}/{repo}");
-                if let Some(ref_or_rev) = ref_or_rev {
-                    write!(f, "/{ref_or_rev}");
-                }
-                Ok(())
-            }
-            FlakeRefType::Indirect { id, ref_or_rev } => {
-                if let Some(ref_or_rev) = ref_or_rev {
-                    write!(f, "{id}/{ref_or_rev}")
-                } else {
-                    write!(f, "{id}")
-                }
-            }
-            FlakeRefType::Mercurial { url, r#type } => todo!(),
-            FlakeRefType::Path { path } => todo!(),
-            // TODO: alternate tarball representation
-            FlakeRefType::Tarball { url, r#type } => {
-                write!(f, "file:{url}")
-            }
-            FlakeRefType::None => todo!(),
-        }
-    }
-}
-
-#[derive(Debug, Default, Clone, Serialize, Deserialize, PartialEq, Eq)]
-pub enum UrlType {
-    #[default]
-    None,
-    Https,
-    Ssh,
-    File,
-}
-
-impl TryFrom<&str> for UrlType {
-    type Error = NixUriError;
-
-    fn try_from(value: &str) -> Result<Self, Self::Error> {
-        use UrlType::*;
-        match value {
-            "" => Ok(None),
-            "https" => Ok(Https),
-            "ssh" => Ok(Ssh),
-            "file" => Ok(File),
-            err => Err(NixUriError::UnknownUrlType(err.into())),
-        }
-    }
-}
-
-impl Display for UrlType {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            UrlType::None => write!(f, "No Url Type Specified"),
-            UrlType::Https => write!(f, "https"),
-            UrlType::Ssh => write!(f, "ssh"),
-            UrlType::File => write!(f, "file"),
-        }
-    }
-}
-
-impl FlakeRefType {
-    /// Parse type specific information, returns the [`FlakeRefType`]
-    /// and the unparsed input
-    pub fn parse_type(input: &str) -> NixUriResult<FlakeRefType> {
-        use nom::sequence::separated_pair;
-        let (_, maybe_explicit_type) = opt(separated_pair(
-            take_until::<&str, &str, (&str, nom::error::ErrorKind)>(":"),
-            tag(":"),
-            rest,
-        ))(input)?;
-        if let Some((flake_ref_type_str, input)) = maybe_explicit_type {
-            match flake_ref_type_str {
-                "github" | "gitlab" | "sourcehut" => {
-                    let (input, owner_and_repo_or_ref) = parse_owner_repo_ref(input)?;
-                    let mut parsed_iter = owner_and_repo_or_ref.map(|s| s.to_string());
-                    let er_fn = |st: &str| {
-                        NixUriError::MissingTypeParameter(flake_ref_type_str.into(), st.to_string())
-                    };
-                    let owner = parsed_iter.next().ok_or(er_fn("owner"))?;
-                    let repo = parsed_iter.next().ok_or(er_fn("repo"))?;
-                    let ref_or_rev = parsed_iter.next();
-                    let platform = match flake_ref_type_str {
-                        "github" => GitForge::GitHub,
-                        "gitlab" => GitForge::GitLab,
-                        "sourcehut" => GitForge::SourceHut,
-                        _ => unreachable!(),
-                    };
-                    let res = FlakeRefType::GitForge {
-                        platform,
-                        owner,
-                        repo,
-                        ref_or_rev,
-                    };
-                    Ok(res)
-                }
-                "path" => {
-                    // TODO: check if path is an absolute path, if not error
-                    let path = Path::new(input);
-                    // TODO: make this check configurable for cli usage
-                    if !path.is_absolute() || input.contains(']') || input.contains('[') {
-                        return Err(NixUriError::NotAbsolute(input.into()));
-                    }
-                    if input.contains('#') || input.contains('?') {
-                        return Err(NixUriError::PathCharacter(input.into()));
-                    }
-                    let flake_ref_type = FlakeRefType::Path { path: input.into() };
-                    Ok(flake_ref_type)
-                }
-
-                _ => {
-                    if flake_ref_type_str.starts_with("git+") {
-                        let url_type = parse_url_type(flake_ref_type_str)?;
-                        let (input, _tag) =
-                            opt(tag::<&str, &str, (&str, nom::error::ErrorKind)>("//"))(input)?;
-                        let flake_ref_type = FlakeRefType::Git {
-                            url: input.into(),
-                            r#type: url_type,
-                        };
-                        Ok(flake_ref_type)
-                    } else if flake_ref_type_str.starts_with("hg+") {
-                        let url_type = parse_url_type(flake_ref_type_str)?;
-                        let (input, _tag) =
-                            tag::<&str, &str, (&str, nom::error::ErrorKind)>("//")(input)?;
-                        let flake_ref_type = FlakeRefType::Mercurial {
-                            url: input.into(),
-                            r#type: url_type,
-                        };
-                        Ok(flake_ref_type)
-                    } else {
-                        Err(NixUriError::UnknownUriType(flake_ref_type_str.into()))
-                    }
-                }
-            }
-        } else {
-            // Implicit types can be paths, indirect flake_refs, or uri's.
-            if input.starts_with('/') || input == "." {
-                let flake_ref_type = FlakeRefType::Path { path: input.into() };
-                let path = Path::new(input);
-                // TODO: make this check configurable for cli usage
-                if !path.is_absolute()
-                    || input.contains(']')
-                    || input.contains('[')
-                    || !input.is_ascii()
-                {
-                    return Err(NixUriError::NotAbsolute(input.into()));
-                }
-                if input.contains('#') || input.contains('?') {
-                    return Err(NixUriError::PathCharacter(input.into()));
-                }
-                return Ok(flake_ref_type);
-            }
-            //TODO: parse uri
-            let (input, mut owner_and_repo_or_ref) = parse_owner_repo_ref(input)?;
-            if let Some(id) = owner_and_repo_or_ref.next() {
-                if !id
-                    .chars()
-                    .all(|c| c.is_ascii_alphabetic() || c.is_control())
-                    || id.is_empty()
-                {
-                    return Err(NixUriError::InvalidUrl(input.into()));
-                }
-                let flake_ref_type = FlakeRefType::Indirect {
-                    id: id.to_string(),
-                    ref_or_rev: owner_and_repo_or_ref.next().map(|s| s.to_string()),
-                };
-                Ok(flake_ref_type)
-            } else {
-                let (_input, mut owner_and_repo_or_ref) = parse_owner_repo_ref(input)?;
-                let id = if let Some(id) = owner_and_repo_or_ref.next() {
-                    id
-                } else {
-                    input
-                };
-                if !id.chars().all(|c| c.is_ascii_alphabetic()) || id.is_empty() {
-                    return Err(NixUriError::InvalidUrl(input.into()));
-                }
-                Ok(FlakeRefType::Indirect {
-                    id: id.to_string(),
-                    ref_or_rev: owner_and_repo_or_ref.next().map(|s| s.to_string()),
-                })
-            }
-        }
-    }
-    /// Extract a common identifier from it's [`FlakeRefType`] variant.
-    pub(crate) fn get_id(&self) -> Option<String> {
-        match self {
-            FlakeRefType::GitForge { repo, .. } => Some(repo.to_string()),
-            FlakeRefType::File { .. }
-            | FlakeRefType::Git { .. }
-            | FlakeRefType::Tarball { .. }
-            | FlakeRefType::None
-            | FlakeRefType::Indirect { .. }
-            | FlakeRefType::Mercurial { .. }
-            | FlakeRefType::Path { .. } => None,
-        }
-    }
-    pub fn get_repo(&self) -> Option<String> {
-        match self {
-            FlakeRefType::GitForge { repo, .. } => Some(repo.into()),
-            // TODO: return a proper error, if ref_or_rev is tried to be specified
-            FlakeRefType::Mercurial { .. }
-            | FlakeRefType::Path { .. }
-            | FlakeRefType::Indirect { .. }
-            | FlakeRefType::Tarball { .. }
-            | FlakeRefType::File { .. }
-            | FlakeRefType::Git { .. }
-            | FlakeRefType::None => None,
-        }
-    }
-    pub fn get_owner(&self) -> Option<String> {
-        match self {
-            FlakeRefType::GitForge { owner, .. } => Some(owner.into()),
-            // TODO: return a proper error, if ref_or_rev is tried to be specified
-            FlakeRefType::Mercurial { .. }
-            | FlakeRefType::Path { .. }
-            | FlakeRefType::Indirect { .. }
-            | FlakeRefType::Tarball { .. }
-            | FlakeRefType::File { .. }
-            | FlakeRefType::Git { .. }
-            | FlakeRefType::None => None,
-        }
-    }
-    pub fn ref_or_rev(&mut self, ref_or_rev_alt: Option<String>) -> Result<(), NixUriError> {
-        match self {
-            FlakeRefType::GitForge { ref_or_rev, .. }
-            | FlakeRefType::Indirect { ref_or_rev, .. } => {
-                *ref_or_rev = ref_or_rev_alt;
-            }
-            // TODO: return a proper error, if ref_or_rev is tried to be specified
-            FlakeRefType::Mercurial { .. }
-            | FlakeRefType::Path { .. }
-            | FlakeRefType::Tarball { .. }
-            | FlakeRefType::File { .. }
-            | FlakeRefType::Git { .. }
-            | FlakeRefType::None => todo!(),
-        }
-        Ok(())
-    }
-}
 
 impl TryFrom<&str> for FlakeRef {
     type Error = NixUriError;
@@ -579,6 +113,13 @@ impl std::str::FromStr for FlakeRef {
 mod tests {
     use super::*;
     use crate::parser::{parse_nix_uri, parse_params};
+
+    #[test]
+    #[ignore]
+    fn parse_type() {
+        let uri = "github:nixos/nixpkgs";
+        assert_eq!(uri, "nixos/nixpkgs");
+    }
 
     #[test]
     fn parse_simple_uri() {
